remove Regexp::Common
I profiled the code to see why it was slow.  The regexps in the common library are very comprehensive but are too slow for us.  Instead of them we are doing a loose match on what looks like an IP addr.

Once the address is found with regex it's parsed with NetAddr::IP, so that should take care of any validation failures.